making export standalone
